+++ internal/app/services/shortener.go
@@ -2,8 +2,8 @@ package services
 
 import (
 	"github.com/ofstudio/go-shortener/internal/app/config"
-	"github.com/ofstudio/go-shortener/internal/storage"
 	"github.com/ofstudio/go-shortener/pkg/shortid"
+	"github.com/ofstudio/go-shortener/pkg/storage"
 	"net/url"
 )
