+++ internal/handlers/api_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"github.com/ofstudio/go-shortener/internal/app/config"
 	"github.com/ofstudio/go-shortener/internal/app/services"
-	"github.com/ofstudio/go-shortener/internal/storage"
+	"github.com/ofstudio/go-shortener/pkg/storage"
 	"github.com/stretchr/testify/require"
 	"net/http"
 	"net/http/httptest"
