+++ internal/handlers/shortener_test.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"github.com/ofstudio/go-shortener/internal/app/config"
 	"github.com/ofstudio/go-shortener/internal/app/services"
-	"github.com/ofstudio/go-shortener/internal/storage"
+	"github.com/ofstudio/go-shortener/pkg/storage"
 	"github.com/stretchr/testify/require"
 	"io"
 	"io/ioutil"
