+++ pkg/storage/interface.go
