+++ pkg/storage/memory.go
